refactor: broke down executor.rs into smaller files each containing functions related to one command
