+++ src/commands/mod.rs
@@ -1,0 +1,4 @@
+pub mod new_project;
+pub mod install_deps;
+pub mod run;
+pub mod doc_gen;
